Added some sanity to voronoi.py's interface.
Previously voronoi.py expected objects with x and y attributes as input,
but emitted 2-tuples. This was annoying because clients would need to
deal with two incompatible representations of 2D points. With this
change the input is expected to be 2-tuples, which is pretty much the
defacto standard for representing 2D points in Python.