Completely rewrote git-multidiff.
Now instead of running git-difftool n+1 times we run it once. The old
implementation used the files created by git-difftool directly, but
since it deletes the temporary files as soon as it completes, we
effectively had to keep it running to keep the files alive. This meant
having O(n) pipes open. It also turned out that it didn't work correctly
with -M (show moves/renames) as the --name-only flag behaves...
inconsistently in this mode.

The new implementation has a helper tool that copies (or hardlinks, when
possible) the files that appear to be temporary. This means we only run
git-difftool once, and we don't rely on the (flaky) --name-only flag so
-M actually works as expected. (Well, mostly. Currently we're copying
/dev/null, which is kind of weird.)