Fixed bug with handling duplicate filenames.
When the same filename appeared in the argument list multiple times,
things would go haywire because we assumed a 1-to-1 relationship between
arguments and buffers, but vim doesn't create multiple buffers for the
same file.

This particularly happened multiple files were added and/or deleted and git-multidiff
was used. eg:

    git multidiff --cached

could result in vim being invoked with:

    foo /dev/null /dev/null bar

(in this example, foo was deleted and bar was added)

This would create 3 buffers: foo, /dev/null, and bar.

This change iterates over arguments instead of buffers. (This means the
single /dev/null buffer will appear in multiple tabs.)