Fixed git-multidiff's outside of repo root behaviour
git-multidiff invokes the diff from the directory in which git-multidiff
was invoked, but _git-multidiff-helper is invoked from the repo's root.
Because these may not be the same, we need to adjust relative filenames
to be absolute, or better yet, to be relative to the directory in which
the diff will run.